tests/local_tests/test_db.py
import os
import psycopg2
from dotenv import load_dotenv

# Load environment variables from .env
load_dotenv()

# Extract credentials
db_config = {
    "dbname":   os.getenv("DB_NAME"),
    "user":     os.getenv("DB_USER"),
    "password": os.getenv("DB_PASSWORD"),
    "host":     os.getenv("DB_HOST", "localhost"),
    "port":     os.getenv("DB_PORT", "5432"),
}

try:
    # Try to connect
    conn = psycopg2.connect(**db_config)
    cursor = conn.cursor()
    cursor.execute("SELECT version();")
    version = cursor.fetchone()
    print("✅ Connected to database successfully.")
    print("PostgreSQL version:", version[0])
    cursor.close()
    conn.close()

except Exception as e:
    print("❌ Failed to connect to the database:")
    print(str(e))
