requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
typing-extensions
sqlalchemy
asyncpg
greenlet
psycopg2-binary
python-dotenv
apache-beam[gcp]
playwright>=1.30
pytest
pytest-asyncio
httpx
boto3>=1.34.0
botocore>=1.34.0
